+++ plugin/web/extensions/sidebar/core/types.ts
@@ -1,12 +1,3 @@
-type actionType =
-  | "updateOverrides"
-  | "screenshot"
-  | "screenshot-save"
-  | "modal-open"
-  | "modal-close";
-
-export type PostMessageProps = { action: actionType; payload?: any };
-
 export type ReearthApi = {
   default?: {
     camera?: Camera;
