+++ plugin/web/types.ts
@@ -1,0 +1,7 @@
+type actionType =
+  | "updateOverrides"
+  | "screenshot"
+  | "screenshot-save"
+  | "modal-open"
+  | "modal-close";
+export type PostMessageProps = { action: actionType; payload?: any };
